+++ src/schemathesis/specs/openapi/adapter/v3_0.py
@@ -16,15 +16,19 @@ from schemathesis.specs.openapi.adapter.protocol import (
     ExtractParameterSchema,
     ExtractRawResponseSchema,
     ExtractResponseSchema,
+    ExtractSchemaForMediaType,
     ExtractSecurityParameters,
     GetBasePath,
     GetDefaultMediaTypes,
+    GetDefaultResponseMediaType,
     GetParameterSerializer,
     GetRequestPayloadContentTypes,
     GetResponseContentTypes,
     IterParameters,
     IterResponseExamples,
     PrepareMultipart,
+    PrepareResponseMediaTypeSchema,
+    ResolveResponseMediaType,
     ValidateSchema,
 )
 
@@ -37,6 +41,10 @@ examples_container_keyword = "examples"
 extract_parameter_schema: ExtractParameterSchema = parameters.extract_parameter_schema_v3
 extract_raw_response_schema: ExtractRawResponseSchema = responses.extract_raw_response_schema_v3
 extract_response_schema: ExtractResponseSchema = responses.extract_response_schema_v3
+prepare_response_media_type_schema: PrepareResponseMediaTypeSchema = responses.prepare_response_media_type_schema
+get_default_response_media_type: GetDefaultResponseMediaType = responses.get_default_response_media_type_v3
+resolve_response_media_type: ResolveResponseMediaType = responses.resolve_response_media_type_v3
+extract_schema_for_media_type: ExtractSchemaForMediaType = responses.extract_schema_for_media_type_v3
 extract_header_schema: ExtractHeaderSchema = responses.extract_header_schema_v3
 iter_parameters: IterParameters = parameters.iter_parameters_v3
 build_path_parameter: BuildPathParameter = parameters.build_path_parameter_v3_0